Issue #1938 Migrate synchronization of metadata with categorical attributes to MetadataManager
