+++ api/src/main/java/com/epam/pipeline/controller/metadata/MetadataController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2019 EPAM Systems, Inc. (https://www.epam.com/)
+ * Copyright 2017-2021 EPAM Systems, Inc. (https://www.epam.com/)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -222,4 +222,15 @@ public class MetadataController extends AbstractRestController {
         return Result.success(metadataApiService.searchMetadataByClassAndKeyValue(entityClass, key, value));
     }
 
+    @PostMapping("/metadata/sync/categoricalAttributes")
+    @ApiOperation(
+        value = "Fill in categorical attributes table based on a current metadata.",
+        notes = "Fill in categorical attributes table based on a current metadata.",
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiResponses(
+        value = {@ApiResponse(code = HTTP_STATUS_OK, message = API_STATUS_DESCRIPTION)
+        })
+    public void syncCategoricalAttributesWithMetadata() {
+        metadataApiService.syncWithCategoricalAttributes();
+    }
 }
