+++ api/src/test/java/com/epam/pipeline/manager/metadata/MetadataManagerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2019 EPAM Systems, Inc. (https://www.epam.com/)
+ * Copyright 2017-2021 EPAM Systems, Inc. (https://www.epam.com/)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,26 +18,49 @@ package com.epam.pipeline.manager.metadata;
 
 import com.epam.pipeline.AbstractSpringTest;
 import com.epam.pipeline.controller.vo.EntityVO;
+import com.epam.pipeline.controller.vo.MetadataVO;
+import com.epam.pipeline.entity.metadata.CategoricalAttribute;
+import com.epam.pipeline.entity.metadata.CategoricalAttributeValue;
 import com.epam.pipeline.entity.metadata.MetadataEntry;
 import com.epam.pipeline.entity.metadata.PipeConfValue;
+import com.epam.pipeline.entity.preference.Preference;
 import com.epam.pipeline.entity.security.acl.AclClass;
+import com.epam.pipeline.entity.user.PipelineUser;
 import com.epam.pipeline.exception.MetadataReadingException;
+import com.epam.pipeline.manager.preference.PreferenceManager;
+import com.epam.pipeline.manager.preference.SystemPreferences;
+import com.epam.pipeline.manager.user.UserManager;
 import org.apache.commons.io.IOUtils;
+import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 public class MetadataManagerTest extends AbstractSpringTest {
 
     @Autowired
     private MetadataManager metadataManager;
+
+    @Autowired
+    private PreferenceManager preferenceManager;
+
+    @Autowired
+    private CategoricalAttributeManager categoricalAttributeManager;
+
+    @Autowired
+    private UserManager userManager;
 
     private Map<String, PipeConfValue> expectedData = new HashMap<>();
     private EntityVO entityVO = new EntityVO(1L, AclClass.PIPELINE);
@@ -52,6 +75,8 @@ public class MetadataManagerTest extends AbstractSpringTest {
     private static final String CSV_HEADER = "Key,Value,Type";
     private static final String TSV_FILE_NAME = "test_file.tsv";
     private static final String CSV_FILE_NAME = "test_file.csv";
+    private static final String SENSITIVE_KEY = "sensitive_metadata_key";
+    private static final String TEST_USER = "TEST_USER";
 
     @Before
     public void setup() {
@@ -109,6 +134,40 @@ public class MetadataManagerTest extends AbstractSpringTest {
                 VALUE_1 + "," + KEY_1 + "\n" +
                 VALUE_2 + "," + KEY_2 + "\n";
         assertFileContent(exampleContent, CSV_FILE_NAME);
+    }
+
+    @Test
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    public void syncWithCategoricalAttributes() {
+        preferenceManager.update(Collections.singletonList(new Preference(
+            SystemPreferences.MISC_METADATA_SENSITIVE_KEYS.getKey(),
+            String.format("[\"%s\"]", SENSITIVE_KEY))));
+        Assert.assertEquals(0, categoricalAttributeManager.loadAll().size());
+
+        final PipelineUser testUser = userManager
+            .createUser(TEST_USER, Collections.emptyList(), Collections.emptyList(), Collections.emptyMap(), null);
+        final EntityVO entityVO = new EntityVO(testUser.getId(), AclClass.PIPELINE_USER);
+        final Map<String, PipeConfValue> data = new HashMap<>();
+        data.put(KEY_1, new PipeConfValue(TYPE, VALUE_1));
+        data.put(KEY_2, new PipeConfValue(TYPE, VALUE_2));
+        data.put(SENSITIVE_KEY, new PipeConfValue(TYPE, VALUE_2));
+        final MetadataVO metadataVO = new MetadataVO();
+        metadataVO.setEntity(entityVO);
+        metadataVO.setData(data);
+        metadataManager.updateMetadataItem(metadataVO);
+        metadataManager.syncWithCategoricalAttributes();
+
+        final Map<String, List<String>> categoricalAttributesAfterSync = categoricalAttributeManager.loadAll().stream()
+            .collect(Collectors.toMap(CategoricalAttribute::getName,
+                attribute -> attribute.getValues().stream()
+                                          .map(CategoricalAttributeValue::getValue)
+                                          .collect(Collectors.toList())));
+        Assert.assertEquals(2, categoricalAttributesAfterSync.size());
+        Assert.assertThat(categoricalAttributesAfterSync.get(KEY_1),
+                          CoreMatchers.is(Collections.singletonList(VALUE_1)));
+        Assert.assertThat(categoricalAttributesAfterSync.get(KEY_2),
+                   CoreMatchers.is(Collections.singletonList(VALUE_2)));
+        Assert.assertFalse(categoricalAttributesAfterSync.containsKey(SENSITIVE_KEY));
     }
 
     @Test(expected = MetadataReadingException.class)
